Lisää tehtävät 4.20.-4.23.

+++ osa4/controllers/blogs.js
@@ -6,14 +6,7 @@ const Blog = require('../models/blog')
 const { response } = require('../app')
 const User = require('../models/user')
 const jwt = require('jsonwebtoken')
-
-const getTokenFrom = request => {
-  const authorization = request.get('authorization')
-  if (authorization && authorization.startsWith('Bearer ')) {
-    return authorization.replace('Bearer ', '')
-  }
-  return null
-}
+const { userExtractor } = require('../utils/middleware')
 
 // hakee blogit tietokannasta
 blogsRouter.get('/', async (request, response) => {
@@ -33,15 +26,13 @@ blogsRouter.get('/:id', async (request, response) => {
 })
 
 // tallentaa blogit tietokantaan
-blogsRouter.post('/', async (request, response) => {
+blogsRouter.post('/', userExtractor, async (request, response) => {
   const body = request.body
+  const user = request.user
 
-  const decodedToken = jwt.verify(getTokenFrom(request), process.env.SECRET)
-  if (!decodedToken.id) {
+  if (!user) {
     return response.status(401).json({ error: 'token invalid' })
   }
-
-  const user = await User.findById(decodedToken.id)
 
   /*
   if (!user) {
@@ -52,10 +43,6 @@ blogsRouter.post('/', async (request, response) => {
     }
   }
   */
-
-  if (!user) {
-    return response.status(400).json({ error: 'userId missing or not valid' })
-  }
 
   const blog = new Blog({
     title: body.title,
@@ -76,21 +63,54 @@ blogsRouter.post('/', async (request, response) => {
 
   const populatedBlog = await savedBlog.populate('user', { username: 1, name: 1 })
 
-  // Muutettu populatedBlogiin
+  // Muutettu populatedBlogiin (oli savedBlog)
   response.status(201).json(populatedBlog)
 })
 
 // poistaa blogin tietokannasta
-blogsRouter.delete('/:id', async (request, response) => {
+blogsRouter.delete('/:id', userExtractor, async (request, response) => {
+  const user = request.user
+
+  if (!user) {
+    return response.status(401).json({ error: 'token invalid' })
+  }
+
+  const blog = await Blog.findById(request.params.id)
+  if (!blog) {
+    return response.status(404).json({ error: 'blog not found'})
+  }
+
+  if (blog.user.toString() !== user._id.toString()) {
+    return response.status(401).json({ error: 'blog can only be deleted by the creator' })
+  }
+
   await Blog.findByIdAndDelete(request.params.id)
   response.status(204).end()
 })
 
 // blogin muokkaaminen
-blogsRouter.put('/:id', async (request, response) => {
+blogsRouter.put('/:id', userExtractor, async (request, response) => {
   const body = request.body
+  const user = request.user
 
-  const editedBlog = await Blog.findByIdAndUpdate(request.params.id, body, { new: true })
+  if (!user) {
+    return response.status(401).json({ error: 'token invalid' })
+  }
+
+  const blog = await Blog.findById(request.params.id)
+  if (!blog) {
+    return response.status(404).json({ error: 'blog not found' })
+  }
+
+  if (blog.user.toString() !== user._id.toString()) {
+    return response.status(401).json({ error: 'blog can only be edited by the creator' })
+  }
+
+  const editedBlog = await Blog.findByIdAndUpdate(
+    request.params.id,
+    body,
+    { new: true }
+    ).populate('user', { username: 1, name: 1 })
 
   if (editedBlog) {
     response.json(editedBlog)
